+++ junos/provider.go
@@ -94,6 +94,7 @@ func Provider() *schema.Provider {
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"junos_interface":          dataSourceInterface(),
+			"junos_interface_physical": dataSourceInterfacePhysical(),
 			"junos_system_information": dataSourceSystemInformation(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
@@ -104,6 +105,7 @@ func Provider() *schema.Provider {
 			"junos_bgp_neighbor":                                         resourceBgpNeighbor(),
 			"junos_firewall_filter":                                      resourceFirewallFilter(),
 			"junos_firewall_policer":                                     resourceFirewallPolicer(),
+			"junos_interface_physical":                                   resourceInterfacePhysical(),
 			"junos_interface_st0_unit":                                   resourceInterfaceSt0Unit(),
 			"junos_interface":                                            resourceInterface(),
 			"junos_ospf_area":                                            resourceOspfArea(),